Keep moving schema agnostic code out of kernel
